# backend/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking Management System API - Full Documentation",
        "description": "Comprehensive API documentation for the Vehicle Parking Management System backend.",
        "version": "1.0.0",
        "contact": {
            "name": "Parking System Support",
            "email": "support@parkingsystem.example.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Development Server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/vehicles": {
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Add a new vehicle",
                "operationId": "addVehicle",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateVehicleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle added successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "vehicle": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "List my vehicles",
                "operationId": "listMyVehicles",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PageQueryParam"
                    },
                    {
                        "$ref": "#/components/parameters/LimitQueryParam"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by plate number or vehicle type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Field to sort vehicles by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "plate_number",
                                "vehicle_type",
                                "size",
                                "created_at"
                            ],
                            "default": "created_at"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/OrderQueryParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of the user's vehicles.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/vehicles/{id}": {
            "parameters": [
                {
                    "$ref": "#/components/parameters/IdPathParamUuid"
                }
            ],
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get my vehicle by ID",
                "operationId": "getMyVehicleById",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update my vehicle by ID",
                "operationId": "updateMyVehicle",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVehicleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "vehicle": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Delete my vehicle by ID",
                "operationId": "deleteMyVehicle",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error description"
                    }
                }
            },
            "Pagination": {
                "type": "object",
                "properties": {
                    "currentPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    },
                    "itemsPerPage": {
                        "type": "integer"
                    }
                }
            },
...
            "VehicleTypeEnum": {
                "type": "string",
                "enum": [
                    "CAR",
                    "MOTORCYCLE",
                    "TRUCK",
                    "BICYCLE"
                ]
            },
            "VehicleSizeEnum": {
                "type": "string",
                "enum": [
                    "SMALL",
                    "MEDIUM",
                    "LARGE",
                    "EXTRA_LARGE"
                ]
            },
...
            "Vehicle": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "plate_number": {
                        "type": "string"
                    },
                    "vehicle_type": {
                        "$ref": "#/components/schemas/VehicleTypeEnum"
                    },
                    "size": {
                        "$ref": "#/components/schemas/VehicleSizeEnum"
                    },
                    "other_attributes": {
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CreateVehicleRequest": {
                "type": "object",
                "required": [
                    "plate_number",
                    "vehicle_type",
                    "size"
                ],
                "properties": {
                    "plate_number": {
                        "type": "string"
                    },
                    "vehicle_type": {
                        "$ref": "#/components/schemas/VehicleTypeEnum"
                    },
                    "size": {
                        "$ref": "#/components/schemas/VehicleSizeEnum"
                    },
                    "other_attributes": {
                        "type": "object",
                        "nullable": true
                    }
                }
            },
            "UpdateVehicleRequest": {
                "type": "object",
                "properties": {
                    "plate_number": {
                        "type": "string"
                    },
                    "vehicle_type": {
                        "$ref": "#/components/schemas/VehicleTypeEnum"
                    },
                    "size": {
                        "$ref": "#/components/schemas/VehicleSizeEnum"
                    },
                    "other_attributes": {
                        "type": "object",
                        "nullable": true
                    }
                }
            },
            "VehicleListResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Vehicle"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                }
            },
...
        },
        "parameters": {
            "IdPathParamUuid": {
                "name": "id",
                "in": "path",
                "required": true,
                "description": "UUID Identifier of the resource",
                "schema": {
                    "type": "string",
                    "format": "uuid"
                }
            },
            "PageQueryParam": {
                "name": "page",
                "in": "query",
                "description": "Page number",
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "LimitQueryParam": {
                "name": "limit",
                "in": "query",
                "description": "Items per page",
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
            "OrderQueryParam": {
                "name": "order",
                "in": "query",
                "description": "Sort order",
                "schema": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ],
                    "default": "desc"
                }
            },
...
        },
        "responses": {
            "UnauthorizedError": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "ForbiddenError": {
                "description": "Forbidden",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "NotFoundError": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "BadRequestError": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            }
        }
    }
}